requirements.txt
lxml==5.2.2
Markdown==3.6
python-docx==1.1.2
typing_extensions==4.12.2
